Moved render methods to render class
