+++ includes/Frontend.php
@@ -89,8 +89,7 @@ class Frontend {
 			$post_template_data = $post_instance->get_template_data();
 
 			if ( ! empty( $post_template_data ) ) {
-				// Register content area
-				$this->register_area( 'content', $post_template_data );
+				Plugin::$instance->renderer->register_area( 'content', $post_template_data );
 
 				// Add content filters
 				add_filter( 'the_content', [ $this, 'add_pagebuilder_content' ], self::CONTENT_FILTER_PRIORITY );
@@ -101,7 +100,7 @@ class Frontend {
 		}
 
 		// Instantiate all elements that are present on the current page
-		$this->prepare_areas_for_render();
+		Plugin::$instance->renderer->prepare_areas_for_render();
 	}
 
 	public function remove_content_filter( $content = '' ) {
@@ -122,7 +121,7 @@ class Frontend {
 	public function add_pagebuilder_content( $content ) {
 		$this->restore_content_filters();
 
-		$pb_content = $this->get_content();
+		$pb_content = Plugin::$instance->renderer->get_content();
 		if ( ! empty( $pb_content ) ) {
 			$content = $pb_content;
 			// Remove filters that may affect content
@@ -154,23 +153,6 @@ class Frontend {
 		add_filter( 'the_content', 'wptexturize' );
 	}
 
-	private function get_content() {
-		ob_start();
-		$this->render_area( 'content' );
-		return ob_get_clean();
-	}
-
-	/**
-	 * Get Elements Instances
-	 *
-	 * Returns an arary containing all element instances for the current page
-	 *
-	 * @return array
-	 */
-	public function get_elements_instances() {
-		return $this->instantiated_elements;
-	}
-
 	public function enqueue_styles() {
 		// Trigger action before load styles
 		do_action( 'zionbuilder/frontend/before_load_styles', $this );
@@ -193,132 +175,5 @@ class Frontend {
 		do_action( 'zionbuilder/frontend/before_load_scripts', $this );
 
 		do_action( 'zionbuilder/frontend/after_load_scripts', $this );
-	}
-
-	/**
-	 * Get Registered Areas
-	 *
-	 * Returns an array containing all registered areas and their data
-	 *
-	 * @return array
-	 */
-	public function get_registered_areas() {
-		return $this->registered_areas;
-	}
-
-
-	/**
-	 * Get Content For Area
-	 *
-	 * Returns the area content data for the given area id
-	 *
-	 * @param string $area_id
-	 *
-	 * @return array
-	 */
-	public function get_content_for_area( $area_id ) {
-		if ( ! empty( $this->registered_areas[$area_id] ) && is_array( $this->registered_areas[$area_id] ) ) {
-			return $this->registered_areas[$area_id];
-		}
-
-		return [];
-	}
-
-	/**
-	 * Register Area
-	 *
-	 * Register a new area of elements
-	 *
-	 * @param string $area_name
-	 * @param array  $area_template_data
-	 *
-	 * @return void
-	 */
-	public function register_area( $area_name, $area_template_data ) {
-		$this->registered_areas[$area_name] = $area_template_data;
-	}
-
-
-	/**
-	 * Register Element Instance
-	 *
-	 * Registers the element instances so we can use them
-	 *
-	 * @param array $element_data
-	 *
-	 * @return void
-	 */
-	private function register_element_instance( $element_data ) {
-		$element_instance_with_data = Plugin::$instance->elements_manager->get_element_instance_with_data( $element_data );
-
-		// Don't proceed if we do not have an element instance
-		if ( false === $element_instance_with_data || ! isset( $element_data['uid'] ) ) {
-			return;
-		}
-
-		$this->instantiated_elements[$element_data['uid']] = $element_instance_with_data;
-
-		// Check if element has children
-		$element_children = $element_instance_with_data->get_children();
-
-		// Instantiate all children elements
-		if ( ! empty( $element_children ) && is_array( $element_children ) ) {
-			foreach ( $element_children as $child_element_data ) {
-				$this->register_element_instance( $child_element_data );
-			}
-		}
-	}
-
-
-	/**
-	 * Prepare content for render
-	 *
-	 * Instantiate all elements that should be rendered on the current page
-	 *
-	 * @return void
-	 */
-	public function prepare_areas_for_render() {
-		foreach ( $this->get_registered_areas() as $area_name => $area_template_data ) {
-			if ( is_array( $area_template_data ) ) {
-				foreach ( $area_template_data as $element_data ) {
-					$this->register_element_instance( $element_data );
-				}
-			}
-		}
-	}
-
-	public function render_area( $area_id ) {
-		echo '<div class="zb zb-area-' . esc_attr( $area_id ) . '">';
-		$this->render_children( $this->get_content_for_area( $area_id ) );
-		echo '</div>';
-	}
-
-	/**
-	 * Render Children
-	 *
-	 * Will loop trough all provided elements and will render them
-	 *
-	 * @param array $children Array containing all children elements
-	 *
-	 * @return void
-	 */
-	public function render_children( $children ) {
-		foreach ( $children as $element_id => $element_data ) {
-			$this->render_element( $element_data );
-		}
-	}
-
-	/**
-	 * Render a single element
-	 *
-	 * @param array $element_data
-	 * @param array $extra_data
-	 *
-	 * @return void
-	 */
-	public function render_element( $element_data, $extra_data = [] ) {
-		if ( isset( $element_data['uid'] ) && isset( $this->instantiated_elements[$element_data['uid']] ) ) {
-			$this->instantiated_elements[$element_data['uid']]->render_element( $extra_data );
-		}
 	}
 }
