+++ includes/Scripts.php
@@ -67,6 +67,28 @@ class Scripts {
 
 
 	/**
+	 * Register a script.
+	 *
+	 * Registers the script if $src provided (does NOT overwrite).
+	 *
+	 * @since 1.0.0
+	 *
+	 * @param string           $handle    Name of the script. Should be unique.
+	 * @param string           $src       relative URL of the script, or path of the script relative to the Project root directory.
+	 *                                    Default empty.
+	 * @param string[]         $deps      Optional. An array of registered script handles this script depends on. Default empty array.
+	 * @param string|bool|null $ver       Optional. String specifying script version number, if it has one, which is added to the URL
+	 *                                    as a query string for cache busting purposes. If version is set to false, a version
+	 *                                    number is automatically added equal to current installed WordPress version.
+	 *                                    If set to null, no version is added.
+	 * @param bool             $in_footer Optional. Whether to enqueue the script before </body> instead of in the <head>.
+	 *                                    Default 'false'.
+	 */
+	public function register_script( $handle, $src = '', $deps = [], $ver = false, $in_footer = false ) {
+		wp_register_script( $handle, $this->get_script_url( $src ), $deps, $ver, $in_footer );
+	}
+
+	/**
 	 * Enqueue a script.
 	 *
 	 * Registers the script if $src provided (does NOT overwrite), and enqueues it.
@@ -84,7 +106,7 @@ class Scripts {
 	 * @param bool             $in_footer Optional. Whether to enqueue the script before </body> instead of in the <head>.
 	 *                                    Default 'false'.
 	 */
-	public function enqueue_script( $handle, $src = '', $deps = array(), $ver = false, $in_footer = false ) {
+	public function enqueue_script( $handle, $src = '', $deps = [], $ver = false, $in_footer = false ) {
 		wp_enqueue_script( $handle, $this->get_script_url( $src ), $deps, $ver, $in_footer );
 	}
 
@@ -95,19 +117,19 @@ class Scripts {
 	 *
 	 * @since 1.0.0
 	 *
-	 * @param string           $handle    Name of the script. Should be unique.
-	 * @param string           $src       relative URL of the script, or path of the script relative to the Project root directory.
-	 *                                    Default empty.
-	 * @param string[]         $deps      Optional. An array of registered script handles this script depends on. Default empty array.
-	 * @param string|bool|null $ver       Optional. String specifying script version number, if it has one, which is added to the URL
-	 *                                    as a query string for cache busting purposes. If version is set to false, a version
-	 *                                    number is automatically added equal to current installed WordPress version.
-	 *                                    If set to null, no version is added.
+	 * @param string           $handle Name of the script. Should be unique.
+	 * @param string           $src    relative URL of the script, or path of the script relative to the Project root directory.
+	 *                                 Default empty.
+	 * @param string[]         $deps   Optional. An array of registered script handles this script depends on. Default empty array.
+	 * @param string|bool|null $ver    Optional. String specifying script version number, if it has one, which is added to the URL
+	 *                                 as a query string for cache busting purposes. If version is set to false, a version
+	 *                                 number is automatically added equal to current installed WordPress version.
+	 *                                 If set to null, no version is added.
 	 * @param string           $media  Optional. The media for which this stylesheet has been defined.
 	 *                                 Default 'all'. Accepts media types like 'all', 'print' and 'screen', or media queries like
 	 *                                 '(orientation: portrait)' and '(max-width: 640px)'.
 	 */
-	public function enqueue_style( $handle, $src = '', $deps = array(), $ver = false, $media = 'all' ) {
+	public function enqueue_style( $handle, $src = '', $deps = [], $ver = false, $media = 'all' ) {
 		wp_enqueue_style( $handle, $this->get_script_url( $src ), $deps, $ver, $media );
 	}
